migration, memory: explicit dimm configuration
Migration needs to replicate the dimm configuration on the
remote side.
Since our explicit dimm configuration is still partially
abstracted in a way that we don't need to care about
addresses or slots, we need a way to pass this information
to the remote side.
This "upgrades" the stdin information passing we use for
the spice ticket to a json object which can now contain the
spice ticket as well as qemu's current dimm list obtained
from the 'query-memory-devices' qmp command.