implement the Machine class
This replaces the old config_to_command() and hotplugging
code, providing a class representing a qemu machine with
methods to perform addition and removal of devices.

This class is responsible for placing devices in the
appropriate slots for the underlying virtual hardware,
checking whether the settings are correct, and issuing the
required qemu runtime commands for hotplugging.

The idea is to separate the base machine type (i440fx, q35)
out in a way that allows adding more architectures,
specifically arm64 wth the arm_virt base machine setup.

Additionally this provides a base that may make it possible
in the future to translate running qemu processes to a
configuration usable for live-migrating still-running
machines after otherwise breaking changes to the way
hardware is put together.

Signed-off-by: Wolfgang Bumiller <w.bumiller@proxmox.com>